Super big commit containing 3 things. YOU SHOULD READ THE COMMIT LOG. * Every function related to the interpretted dependency computation are now deprecated and are protected by #ifdef DEPRECATED. * Corrected Grapher to stop calling for deprecated functions. I had to change slightly the precompile.c to accomodate with NULL "data". Untested, might not work. Might also have (minor, even less if we predict branch at compile time) )side effect on performance. We could consider splitting the deprecated for grapher use only instead. * Removed the need to define DISTRIBUTED by hand. DISTRIBUTED is now magically setup by remote_dep.h when needed (aka when some transport, currently only USE_MPI) is defined. When use MPI is not defined, any code protected by DISTRIBUTED is compiled out, as well as many (all) the remote_dep functions, except in DEBUG mode, where calling remote_dep calls stub functions to generate warnings. * Corrected precompile.c: precompile.c should never be compiled with USE_MPI as dpc is a single node utility. Therefore, any DISTRIBUTED ifdef should appear in the generated code, but not in precompile.c * Changed fw_mask to work on uint32, because working on uint8 is bad for performance (alignment problems, cannot go in registers, needs to rely on gcc to do the smart thing, which is not granted, ...).

+++ execution_unit.h
@@ -38,7 +38,7 @@ typedef struct dplasma_execution_unit_t {
     int8_t*  eu_steal_from;
 #endif  /* !defined(DPLASMA_USE_GLOBAL_LIFO) */
 
-    uint8_t* remote_dep_fw_mask;
+    uint32_t* remote_dep_fw_mask;
 } dplasma_execution_unit_t;
 
 struct dplasma_context_t {
